Updates SPI implementation to maintain backwards compatibility of older SSE implementation, which is also kept around. Changes to SseSink to support multiple implementations.
